Refactor PR commit: - Remove Priority and State class to entity - make TudeeExecption and make him extend Execption and make
