--- app/src/main/java/com/qudus/tudee/data/repository/repository.kt
@@ -1,2 +1,0 @@
-package com.qudus.tudee.data.repository
-
