+++ app/src/main/java/com/qudus/tudee/domain/entity/State.kt
@@ -1,4 +1,4 @@
-package com.qudus.tudee.domain.type
+package com.qudus.tudee.domain.entity
 
 enum class State {
     TODO,
