+++ app/src/main/java/com/qudus/tudee/domain/exception/TudeeExecption.kt
@@ -1,0 +1,4 @@
+package com.qudus.tudee.domain.exception
+
+open class TudeeExecption : Exception(){
+}
