+++ app/src/main/java/com/qudus/tudee/ui/designSystem/component/TudeeComponent.kt
@@ -1,4 +1,4 @@
-package com.qudus.tudee.designSystem.component
+package com.qudus.tudee.ui.designSystem.component
 
 import androidx.compose.runtime.Composable
