+++ app/src/main/java/com/qudus/tudee/ui/designSystem/theme/Theme.kt
@@ -1,11 +1,11 @@
-package com.qudus.tudee.designSystem.theme
+package com.qudus.tudee.ui.designSystem.theme
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.ReadOnlyComposable
-import com.qudus.tudee.designSystem.color.LocalTudeeColors
-import com.qudus.tudee.designSystem.color.TudeeColors
-import com.qudus.tudee.designSystem.textStyle.LocalTudeeTextStyle
-import com.qudus.tudee.designSystem.textStyle.TudeeTextStyle
+import com.qudus.tudee.ui.designSystem.color.LocalTudeeColors
+import com.qudus.tudee.ui.designSystem.color.TudeeColors
+import com.qudus.tudee.ui.designSystem.textStyle.LocalTudeeTextStyle
+import com.qudus.tudee.ui.designSystem.textStyle.TudeeTextStyle
 
 object Theme {
     val color: TudeeColors
