+++ app/src/main/java/com/qudus/tudee/ui/designSystem/theme/TudeeTheme.kt
@@ -1,12 +1,12 @@
-package com.qudus.tudee.designSystem.theme
+package com.qudus.tudee.ui.designSystem.theme
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import com.qudus.tudee.designSystem.color.LocalTudeeColors
-import com.qudus.tudee.designSystem.color.darkThemeColor
-import com.qudus.tudee.designSystem.color.lightThemeColor
-import com.qudus.tudee.designSystem.textStyle.LocalTudeeTextStyle
-import com.qudus.tudee.designSystem.textStyle.defaultTextStyle
+import com.qudus.tudee.ui.designSystem.color.LocalTudeeColors
+import com.qudus.tudee.ui.designSystem.color.darkThemeColor
+import com.qudus.tudee.ui.designSystem.color.lightThemeColor
+import com.qudus.tudee.ui.designSystem.textStyle.LocalTudeeTextStyle
+import com.qudus.tudee.ui.designSystem.textStyle.defaultTextStyle
 
 @Composable
 fun TudeeTheme(
